+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.1.1 (unreleased)
 
-- Added `find_best_n_motifs` method
+- Added features and clustering
 - Added more statistics
+- Added `find_best_n_motifs` method
 
 ## 0.1.0 (2020-12-17)