+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.1 (unreleased)
 
-- Added clustering, dimensionality, and features
+- Added clustering, dimensionality, features, and linalg
 - Added more statistics
 - Added `find_best_n_motifs` method