+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 ## 0.1.1 (unreleased)
 
-- Added clustering, dimensionality, features, and linalg
-- Added more statistics
+- Added many more methods
 - Added `find_best_n_motifs` method
 
 ## 0.1.0 (2020-12-17)